-Added the Favourite button -Changed the percentage to a number out of 10 cause it reminds me of rotten tomatoes and rotten tomatoes sucks -changed the app name and icon -changed the colors of progress bars and favourite icons
On the surface these achievements seem subtle, but behind every commit there is a story, a lore, a fight between good and evil.

it all started when the project wasnt working anywhere and then mahmoud who's real name is hackar woke up to this catastrophe. it took hours and hours to fix the problems due to github being a pile of dogshit

I had to work alot yesterday to fix corrupted files in android studio, by using techniques forbidden by google itself.
today my friends we are witnessing one of the biggest accomplishments in human history (if not the biggest).
after hackar and noor struggling for hours to make the favourite button, i have created it in 30 minutes today although i have a final tomorrow. I refused giving up, remembering 2 of my teammates failing to make this button fueled me to fight until the end and win the battle. it turned from a coding task into an attack on our prides. and we fought till the end my brothers.
This achievement will forever be remembered in the computer scientists community.